+++ src/red_struct_filler.c
@@ -1,12 +1,12 @@
 #include "minishell.h"
-
+#define LITERAL_PATH "PATH"
 #define COMMAND_NOT_FOUND "Error: command {0} not found.\n"
 #define PATH_NOT_FOUND "Error: env variable 'PATH' not found.\n"
 
 /*
 ?   (Continuación de la función de abajo)
 */
-char	*new_get_command_path(char *command, t_str **env_list)
+char	*new_get_command_path(char *command)
 {
 	int		idx;
 	char	*str_att;
@@ -14,7 +14,7 @@ char	*new_get_command_path(char *command, t_str **env_list)
 	char	**path_list;
 
 	idx = 0;
-	paths = ft_getenv(env_list, "PATH");
+	paths = ft_getenv(LITERAL_PATH);
 	if (!paths[0])
 	{
 		log_error(PATH_NOT_FOUND, 1);
@@ -40,7 +40,7 @@ char	*new_get_command_path(char *command, t_str **env_list)
 ?   Si le pasas la dirección a un ejecutable, te devuelve esa misma
 ?   dirección malloqueada.
 */
-char	*new_getpath(char *raw_cmd, t_str **env_list)
+char	*new_getpath(char *raw_cmd)
 {
 	int a;
 
@@ -49,13 +49,13 @@ char	*new_getpath(char *raw_cmd, t_str **env_list)
 		a++;
 	if (raw_cmd[a] == '/')							//TODO		This isn't detecting the first '/' properly :(
 	{
-		if (!access(new_get_command_path(raw_cmd, env_list), X_OK))
+		if (!access(new_get_command_path(raw_cmd), X_OK))
 			return (ft_strdup(raw_cmd));
 		else
 			perror("Error:");
 		return (NULL);
 	}
-	return (new_get_command_path(raw_cmd, env_list));
+	return (new_get_command_path(raw_cmd));
 }
 
 /*
@@ -164,7 +164,7 @@ char	**getredirections(char **list)
 	return (ret);
 }
 
-int  put_params_in_struct(char **list, t_str **env_list, t_red **red_list)
+int  put_params_in_struct(char **list, t_red **red_list)
 {
 	int		a;
 	int		items;
@@ -177,7 +177,7 @@ int  put_params_in_struct(char **list, t_str **env_list, t_red **red_list)
 		item_red = lst_red_new();
 		items++;
 		item_red -> params = getparams(&list[a]);
-		item_red -> path = new_getpath(item_red -> params[0], env_list);
+		item_red -> path = new_getpath(item_red -> params[0]);
 		item_red -> redirs = getredirections(&list[a]);
 		item_red -> pip_in = 0;
 		item_red -> pip_out = 1;